_atoi

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -9,7 +9,7 @@
 int _atoi(char *s)
 {
 	int len, i, sign;
-	long int n;
+	int n;
 
 	len = 0;
 	while (s[len] != '\0')
